requirements.txt
Flask
flask-cors
requests
Pillow
APScheduler
python-dotenv
vosk
sounddevice
numpy